+++ app/android/app/src/main/kotlin/com/rukkafolio/rukka_folio/MainActivity.kt
@@ -1,0 +1,5 @@
+package com.rukkafolio.rukka_folio
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity : FlutterActivity()